Don't convert given defs to vals
